ml-model/requirements.txt
tensorflow>=2.20.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0
tqdm>=4.65.0
kaggle>=1.5.0
kagglehub
scikit-learn>=1.3.0
